Separate decorators into a directory and files
