+++ models/decorators/trimmer_decorator.rb
@@ -1,0 +1,5 @@
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name[0..9]
+  end
+end
